# apps/rest-api/docs/swagger.yaml
definitions:
  model.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
info:
  contact:
    email: wahyu@uloy.dev
    name: Uloydev
  description: Kotak Email Internal Rest API
  title: Kotak Email API
  version: "1.0"
paths:
  /users:
    get:
      description: get users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UserResponse'
            type: array
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create a user
      parameters:
      - description: User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      description: delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete a user
      tags:
      - users
    get:
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: Show a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: Update a user
      tags:
      - users
swagger: "2.0"